pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "runpod-langchain"
version = "0.1.0"
description = "LangChain integration for RunPod serverless endpoints"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["runpod", "langchain", "llm", "ai", "ml", "serverless"]
dependencies = [
    "langchain-core>=0.1.0",
    "requests>=2.25.0",
    "pydantic>=2.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/runpod-langchain"
Documentation = "https://github.com/yourusername/runpod-langchain#readme"
Repository = "https://github.com/yourusername/runpod-langchain"
Issues = "https://github.com/yourusername/runpod-langchain/issues"

[tool.setuptools]
packages = ["runpod_langchain"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]